IPfilter refactored to reduce complexity and for better packet-rule matching.

+++ modules/pico_ipfilter.h
@@ -16,12 +16,12 @@ enum filter_action {
     FILTER_COUNT
 };
 
-int pico_ipv4_filter_add(struct pico_device *dev, uint8_t proto,
+uint32_t pico_ipv4_filter_add(struct pico_device *dev, uint8_t proto,
                          struct pico_ip4 *out_addr, struct pico_ip4 *out_addr_netmask, struct pico_ip4 *in_addr,
                          struct pico_ip4 *in_addr_netmask, uint16_t out_port, uint16_t in_port,
                          int8_t priority, uint8_t tos, enum filter_action action);
 
-int pico_ipv4_filter_del(uint8_t filter_id);
+int pico_ipv4_filter_del(uint32_t filter_id);
 
 int ipfilter(struct pico_frame *f);
